pyproject.toml
[project]
name = "shrlnk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "flask>=3.1.0",
    "flask-limiter>=3.9.2",
    "pillow>=11.0.0",
    "requests>=2.32.3",
]
